+++ modules/tinymce/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a new `toolbar_sticky_offset` setting to allow the toolbar to be offset from the top of the page #TINY-7337
 - Added a new `mceFocus` command that focuses the editor. Equivalent to using `editor.focus()` #TINY-7373
 - Added a new `tablecellvalign` toolbar button and menu item for vertical table cell alignment #TINY-7477
+- Added a new `tablecellborderwidth` toolbar button and menu item to change table cell border width #TINY-7478
+- Added a new `tablecellborderstyle` toolbar button and menu item to change table cell border style #TINY-7478
 
 ### Improved
 - Improved the load time of the `fullpage` plugin by using the existing editor schema rather than creating a new one #TINY-6504